+++ CHANGELOG.md
@@ -5,6 +5,8 @@ History
 ------------------
 
 - Shell: Fix binary shims when `-h`/`--help` is given as argument.
+- Core: Add the `info` command which output compacted information about docker containers such as environment 
+  variables, virtual host, exposed ports and binaries.
 
 
 1.0.3 (2020-09-01)